requirements.txt
alembic==1.10.4
aniso8601==9.0.1
attrs==23.1.0
bcrypt==4.0.1
boto3==1.26.121
botocore==1.29.121
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
dnspython==2.3.0
dominate==2.7.0
email-validator==1.1.3
et-xmlfile==1.1.0
Flask==2.0.2
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
flask-bootstrap5==0.1.dev1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-Minify==0.37
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
fpdf==1.7.2
greenlet==2.0.2
gunicorn==20.1.0
htmlmin==0.1.12
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.0.2
jmespath==1.0.1
jsmin==3.0.1
jsonschema==4.17.3
lesscpy==0.15.1
Mako==1.2.4
MarkupSafe==2.1.2
numpy==1.24.3
openpyxl==3.1.2
pkgutil_resolve_name==1.3.10
ply==3.11
psycopg2-binary==2.9.7
pycparser==2.21
PyJWT==2.6.0
pypdf==3.12.2
PyPDF2==3.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.19.2
python-simplexml==0.1.5
pytz==2023.3
rcssmin==1.1.1
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==1.4.29
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
visitor==0.1.3
Werkzeug==2.0.3
WTForms==3.0.0
xxhash==3.2.0
zipp==3.15.0
